chore(plot): move LineSource into bokeh_source.py
